update to the mesh methods, refactor to create elementspace and nodespace
